requirements.txt
numpy
matplotlib
pandas
pyserial
pybaselines
scikit-learn
scipy
streamlit
